Use correct type
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/trunk/libsidplayfp@2570 75748be7-0387-4326-94c7-f41ce0883d2c

+++ src/builders/residfp-builder/residfp/WaveformGenerator.cpp
@@ -140,7 +140,7 @@ void WaveformGenerator::setChipModel(ChipModel chipModel)
     Dac dacBuilder(DAC_BITS);
     dacBuilder.kinkedDac(chipModel);
 
-    const float offset = dacBuilder.getOutput(is6581 ? 0x380 : 0x800);
+    const double offset = dacBuilder.getOutput(is6581 ? 0x380 : 0x800);
 
     for (unsigned int i = 0; i < (1 << DAC_BITS); i++)
     {
